requirements.txt
discord.py 
aiohttp 
python-dotenv 
sentence-transformers 
scikit-learn 
numpy 
cachetools